clean up client.py
